extract Message to its own component ofc
